refactor(Tooltip): TooltipとTooltipPortalをTooltip/配下にまとめる
TooltipPortalはTooltipの実装詳細であり、ControlledTooltipとは
独立したコンポーネントであるため、1階層下にまとめて関係を明確にする。

Tooltip/index.tsは'use client'を持つTooltip.tsxのみをexportするため、
バレルの平坦化による副作用importの混入は発生しない。
ビルド出力でTooltip.jsが境界として保持されることを確認済み。
