+++ packages/smarthr-ui/src/components/Tooltip/Tooltip/index.ts
@@ -1,0 +1,1 @@
+export { Tooltip } from './Tooltip'
